Intermediate-Flask/flask-feedback/requirements.txt
bcrypt==4.0.1
blinker==1.5
Click==8.1.3
Flask==2.0.0
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
flask-wtf-1.0.1
greenlet==2.0.1
itsdangerous==2.1.2
IPython==8.7.0
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
SQLAlchemy==1.4.44
Werkzeug==2.2.2
WTForms==3.0.1